requirements.txt
python-dotenv==1.0.1
bcrypt==4.0.1
python-multipart==0.0.20
fastapi==0.115.6
pydantic[email]==2.10.4
psycopg2-binary==2.9.10
passlib==1.7.4
sqlalchemy==2.0.36
google-cloud-firestore==2.19.0
pyjwt==2.10.1
uvicorn==0.34.0
pydantic==2.10.4